refactor: Created a dedicated file for each function related to handling files
