+++ tests/utils/files/create-files.test.js
@@ -5,7 +5,7 @@ import {
 } from '../../helpers/shared-test-setups/typescript.js';
 import { assertFixture, loadFixture, test } from '../../helpers/testing.js';
 
-test('utils | files > createFiles', function () {
+test('utils | files | create-files', function () {
   const inputProject = {};
 
   const outputProject = {
